Fix double ouput of 'comments powered by Disqus.'
If JavaScript is disabled in a web-browser, the message 'comments powered
by Disqus' is doubled.

This issue occurs because the output of the text is additionally appears in
the <noscript> area for JavaScript.

The text in the <noscript> area is deleted so that the Information
appears correct when JavaScript is disabled.